+++ src/common/CMakeLists.txt
@@ -1,3 +1,4 @@
-add_library(COMMON utils.c distributions.cpp)
+add_library(COMMON utils.c distributions.c distributions_first.cpp)
 set_target_properties(COMMON PROPERTIES OUTPUT_NAME common)
 target_include_directories(COMMON PUBLIC "${PROJECT_SOURCE_DIR}/include/common")
+target_link_libraries(COMMON PRIVATE gsl)
